Address PR review comments

Make OIDC IAM permissions optional

rename feature to distinguish from managed mode

Rework OIDC reconcile to work mgmt side only

Fixes and tidy up

Add OIDC provider flavor

use keyIDFromPublicKey from k8s code

Gate OIDC reconcile behind feature flag.

Revert unneeded awsmachine_controller test changes

fix clusterawsadm template tests

Fix lints & tests

bump pod ID webhook to latest (v0.6.6)

make generate

Re-order reconcile steps.

The Identity provider needs the S3 objects, so let's create them first and avoid flakey failures.

Fixes after rebase

+++ controllers/awscluster_controller.go
@@ -281,11 +281,13 @@ func (r *AWSClusterReconciler) reconcileDelete(ctx context.Context, clusterScope
 		allErrs = append(allErrs, errors.Wrap(err, "error deleting network"))
 	}
 
-	if err := iamService.DeleteOIDCProvider(ctx); err != nil {
-		allErrs = append(allErrs, errors.Wrapf(err, "error deleting OIDC Provider"))
+	if feature.Gates.Enabled(feature.OIDCProviderUnmanagedClusters) {
+		if err := iamService.DeleteOIDCProvider(ctx); err != nil {
+			allErrs = append(allErrs, errors.Wrapf(err, "error deleting OIDC Provider"))
+		}
 	}
 
-	if err := s3Service.DeleteBucket(); err != nil {
+	if err := s3Service.DeleteBucket(ctx); err != nil {
 		allErrs = append(allErrs, errors.Wrapf(err, "error deleting S3 Bucket"))
 	}
 
@@ -386,12 +388,6 @@ func (r *AWSClusterReconciler) reconcileNormal(ctx context.Context, clusterScope
 	}
 	conditions.MarkTrue(awsCluster, infrav1.S3BucketReadyCondition)
 
-	if err := iamService.ReconcileOIDCProvider(context.TODO()); err != nil {
-		conditions.MarkFalse(awsCluster, infrav1.OIDCProviderReadyCondition, infrav1.OIDCProviderReconciliationFailedReason, clusterv1.ConditionSeverityError, "%s", err.Error())
-		clusterScope.Error(err, "failed to reconcile OIDC provider")
-		return reconcile.Result{RequeueAfter: 15 * time.Second}, nil
-	}
-
 	for _, subnet := range clusterScope.Subnets().FilterPrivate() {
 		found := false
 		for _, az := range awsCluster.Status.Network.APIServerELB.AvailabilityZones {
@@ -407,6 +403,14 @@ func (r *AWSClusterReconciler) reconcileNormal(ctx context.Context, clusterScope
 	}
 
 	awsCluster.Status.Ready = true
+
+	if feature.Gates.Enabled(feature.OIDCProviderUnmanagedClusters) {
+		if err := iamService.ReconcileOIDCProvider(context.TODO()); err != nil {
+			conditions.MarkFalse(awsCluster, infrav1.OIDCProviderReadyCondition, infrav1.OIDCProviderReconciliationFailedReason, clusterv1.ConditionSeverityError, "%s", err.Error())
+			clusterScope.Error(err, "failed to reconcile OIDC provider")
+		}
+	}
+
 	return reconcile.Result{}, nil
 }
